nanovc-java-13: Null Pointer when scanning for dangling references. Created a new test file for dangling references specifically.
